% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmvc.data.beta.R
\docType{data}
\name{beta}
\alias{beta}
\title{Example DNA methylation data for dmvc function}
\format{
An object of class \code{"matrix"} with with 1000 rows and 334 columns.
Each row is a CpG, each column is a sample
}
\usage{
data(beta)
}
\description{
DNA methylation data from TCGA-COAD
}
\examples{
data(beta)
}

\keyword{datasets}
