% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmvc.data.covariate.R
\docType{data}
\name{covariate}
\alias{covariate}
\title{Example covariate data for dmvc function}
\format{
An object of class \code{"matrix"} with with 334 rows and 3 vaiables.
\describe{
  \item{group}{Whether the sample is normal or tumor, normal:0, tumor:1}
  \item{gender}{Female or Male}
  \item{age}{age (31--90)}
}
}
\usage{
data(covariate)
}
\description{
Covariate data for 334 TCGA-COAD samples
}
\examples{
data(covariate)
}

\keyword{datasets}
