% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.model.probe.R
\name{fit.model.probe}
\alias{fit.model.probe}
\title{Fit linear model}
\usage{
fit.model.probe(mat, colData)
}
\arguments{
\item{mat}{m value matrix}

\item{colData}{covariate matrix}

}
\value{
A list including a matrix "stats" and a matrix "fitted"
\item{stats}{
A test statistics matrix includes coefficients, standard devidation, z-score
p-value, and an indicator (to list constant covariates to be removed, 
0 means nothing to remove).
}
\item{fitted}{
A matrix to keep fitted values.
}
\item{design}{
A list to keep design matrices.
}
\item{designidx}{
A matrix used to acess design matrices.
}
}
\description{
This functin was used by dmvc function. It loads beta matrix and covariate
matrix to compute statistics and fitted values.
}
\examples{
data(beta)
data(covariate)
fitall=fit.model.probe(mat=beta,colData=covariate)
}
