% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmvc.R
\name{dmvc}
\alias{dmvc}
\title{Perform DMC, DVC, DMVC, and DMVC+ tests for genome-wide CpGs in methylation arrays.}
\usage{
dmvc(beta = beta, covariate = covariate, npermut=100,permut.seed=100,corenumber=1)
}
\arguments{
\item{beta}{Methylation beta value matrix, row for CpGs, column for samples,
it uses sample name as the column name.}

\item{covariate}{covariate matrix, a data frame including all covariates to
adjust for, row for samples, column for covariates, it uses sample name as 
the row name, the matrix must include a "group" column, which is a binary 
indicator (0 for noraml and 1 for tumor) to define two groups of samples 
to compare.}

\item{npermut}{number of permutation for joint tests}

\item{permut.seed}{random seed used by permutation for joint tests}

\item{corenumber}{number of cores used for joint tests; if use corenumber>1, 
a parallel version will be used to speed up the process}
}
\value{
A data frame for the following results.
\item{Mean_normal}{
Mean of beta value for normal samples.
}
\item{Mean_tumor}{
Mean of beta value for tumor samples.
}
\item{Mean_all}{
Mean of beta value for all samples.
}
\item{SD_normal}{
Standard deviation of beta value for normal samples.
}
\item{SD_tumor}{
Standard deviation of beta value for tumor samples.
}
\item{SD_all}{
Standard deviation of beta value for all samples.
}
\item{DMCP}{
p-value from DMC test.
}
\item{DVCP}{
p-value from DVC test.
}
\item{Joint1P}{
Joint test for DMVC (test for hypermethylation and increased
variance in cancer).
}
\item{Joint2P}{
Joint test for DMVC (test for differential methylation in both
direction and increased variance in cancer).
}
\item{LRT1}{
Likelihood ratio test statistics for joint test1.
}
\item{LRT2}{
Likelihood ratio test statistics for joint test2.
}
\item{pho}{
Correlation value.
}
}
\description{
This function is used to compute p values and other test statistics
based on DMC/DVC and joint tests.
}

\references{
James Y. Dai et al. Incorporating increased variability in differential 
methylation and cancer biomarker discovery, 
Genome Biology 2021, submitted.
}
\examples{
data(beta)
data(covariate)
out=dmvc(beta=beta,covariate=covariate)
}

